Separated the epiphany library into multiple files. Added a basic malloc implemenatation
